+++ CHANGELOG.md
@@ -12,12 +12,13 @@ Given version x.y.z
 
 **In case of x == 0 backward compatibility is not guaranteed.**
 
-### [0.0.9-SNAPSHOT]() (26-08-2017)
+### [0.0.9]() (26-08-2017)
 
 * Fixed ability to disable queue processing via thread count
 * Renamed ThreadLifecycleListener#crashedOnPickedTask to ThreadLifecycleListener#crashed
 * Constructors of builder classes are made private
 * Return value of QueueProducer#enqueue changed to primitive
+* Added reenqueue by actor via QueueActorDao#reenqueue
 
 ### [0.0.8]() (24-08-2017)